+++ bosesoundtouchapi/firmware/soundtouchfirmwarerelease.py
@@ -2,17 +2,19 @@
 import xml.etree.ElementTree as xmltree
 
 # our package imports.
+from bosesoundtouchapi.bstutils import export
 from bosesoundtouchapi.soundtoucherror import SoundTouchError
 
 # get smartinspect logger reference; create a new session for this module name.
 import logging
 from smartinspectpython.siauto import SIAuto, SILevel, SISession
-_logsi:SISession = SIAuto.Si.GetSession(__name__)
+_logsi:SISession = SIAuto.Si.GetSession(__package__)
 if (_logsi == None):
-    _logsi = SIAuto.Si.AddSession(__name__, True)
-_logsi.SystemLogger = logging.getLogger(__name__)
+    _logsi = SIAuto.Si.AddSession(__package__, True)
+_logsi.SystemLogger = logging.getLogger(__package__)
 
 
+@export
 class SoundTouchFirmwareRelease:
     """
     A Bose SoundTouch release targets a specific firmware upgrade.
@@ -22,20 +24,20 @@ class SoundTouchFirmwareRelease:
     stored in the XML-Element.
     """
     
-    def __init__(self, revision:str=None, host:str=None, uri:str=None, 
-                 usbUri:str=None, image:dict=None, notesUrl:str=None, 
+    def __init__(self, revision:str=None, httpHost:str=None, urlPath:str=None, 
+                 usbPath:str=None, image:dict=None, notesUrl:str=None, 
                  features:list=None
                  ) -> None:
         """
         Args:
             revision (str):
                 The revision number of the current release.
-            host (str):
-                The hostname of the update provider.
-            uri (str):
+            httpHost (str):
+                The http hostname of the update provider.
+            urlPath (str):
                 The uri part of the full url linked to the downloadable update file.
-            usbUri (str):
-                Another uri which was not usable in any context.
+            usbPath (str):
+                The uri part of the full url linked to the downloadable USB device update file.
             image (dict[str, str]):
                 The main property storing data related to the firmware image.
             notesUrl (str):
@@ -44,13 +46,12 @@ class SoundTouchFirmwareRelease:
                 If there are some features within the release, they are added to this list as a dict.
         """
         self._Features:list = features if features else []
-        self._Host:str = host
+        self._HttpHost:str = httpHost
         self._Image:dict = image if image else {}
         self._NotesUrl:str = notesUrl
         self._Revision:str = revision
-        self._Uri:str = uri
-        self._UsbUri:str = usbUri
-
+        self._UrlPath:str = urlPath
+        self._UsbPath:str = usbPath
 
     def __str__(self) -> str:
         return self.ToString()
@@ -65,11 +66,11 @@ class SoundTouchFirmwareRelease:
 
 
     @property
-    def Host(self) -> str:
+    def HttpHost(self) -> str:
         """
-        Hostname of the update provider.
+        HTTP Hostname of the update provider.
         """
-        return self._Host
+        return self._HttpHost
 
 
     @property
@@ -97,19 +98,19 @@ class SoundTouchFirmwareRelease:
 
 
     @property
-    def Uri(self) -> str:
+    def UrlPath(self) -> str:
         """
         URI part of the full URL linked to the downloadable update file.
         """
-        return self._Uri
+        return self._UrlPath
 
 
     @property
-    def UsbUri(self) -> str:
+    def UsbPath(self) -> str:
         """
-        Another URI which was not usable in any context.
+        The uri part of the full url linked to the downloadable USB device update file.
         """
-        return self._UsbUri
+        return self._UsbPath
 
 
     @staticmethod
@@ -133,20 +134,25 @@ class SoundTouchFirmwareRelease:
         if not element:
             raise SoundTouchError('Expected non null input', logsi=_logsi)
 
+        # process base element details.
         release = SoundTouchFirmwareRelease(
             element.get('REVISION', None),
             element.get('HTTPHOST', None),
             element.get('URLPATH', None),
             element.get('USBPATH', None)
         )
+
+        # process IMAGE element details (if any).
         image = element.find('IMAGE')
         if image is not None:
             release._Image = image.attrib
 
+        # process NOTES element details (if any).
         notes = element.find('NOTES')
         if notes is not None:
             release._NotesUrl = notes.get('URL', None)
 
+        # process FEATURE element details (if any).
         for feature in element.findall('FEATURE'):
             release._Features.append(feature.attrib)
             
@@ -158,5 +164,9 @@ class SoundTouchFirmwareRelease:
         Returns a displayable string representation of the class.
         """
         msg:str = 'SoundTouchFirmwareRelease:'
-        msg = "%s revision='%s'" % (msg, self._Revision)
+        msg = "%s  revision='%s'" % (msg, self._Revision)
+        msg = "%s, host='%s'" % (msg, self._HttpHost)
+        msg = "%s, urlPath='%s'" % (msg, self._UrlPath)
+        msg = "%s, usbPath='%s'" % (msg, self._UsbPath)
+        msg = "%s, notes='%s'" % (msg, self._NotesUrl)
         return msg